Refactored CategoryChart
Moved CategoryChart to its own file
