+++ Carbon/navigation/screens/Progress/ProgressScreen.js
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import { View, SafeAreaView, ScrollView, Dimensions, StyleSheet, Text, Platform } from 'react-native';
 import { Colors } from '../../../colors/Colors';
-import { CategoryChart, KeyFactors } from '../../../components/ChartData';
+import { KeyFactors } from '../../../components/ChartData';
+import { CategoryChart } from '../../../components/CategoryChart';
 import RecordEmission from './RecordEmission';
 import { TouchableOpacity } from 'react-native';
 import { ScreenNames } from '../Main/ScreenNames';
